backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
openai==1.6.1
requests==2.31.0